Moved tests depending on PostgreSQL to 'usesjava8' package as postgresql-embedded requires Java 8.
