+++ src/test/java/org/apache/ibatis/submitted/usesjava8/keycolumn/InsertMapper.java
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.apache.ibatis.submitted.keycolumn;
+package org.apache.ibatis.submitted.usesjava8.keycolumn;
 
 import org.apache.ibatis.annotations.Insert;
 import org.apache.ibatis.annotations.Options;
