+++ src/test/java/org/apache/ibatis/submitted/usesjava8/keycolumn/InsertTest.java
@@ -13,27 +13,27 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.apache.ibatis.submitted.refcursor;
+package org.apache.ibatis.submitted.usesjava8.keycolumn;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
-import java.io.IOException;
 import java.io.Reader;
 import java.nio.file.Paths;
 import java.sql.Connection;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
+import org.apache.ibatis.executor.BatchResult;
 import org.apache.ibatis.io.Resources;
 import org.apache.ibatis.jdbc.ScriptRunner;
+import org.apache.ibatis.session.ExecutorType;
 import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
 import org.apache.ibatis.session.SqlSessionFactoryBuilder;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import ru.yandex.qatools.embed.postgresql.EmbeddedPostgres;
@@ -41,7 +41,7 @@ import ru.yandex.qatools.embed.postgresql.EmbeddedPostgres;
 /**
  * @author Jeff Butler
  */
-public class RefCursorTest {
+public class InsertTest {
 
   private static final EmbeddedPostgres postgres = new EmbeddedPostgres();
 
@@ -50,15 +50,15 @@ public class RefCursorTest {
   @BeforeClass
   public static void setUp() throws Exception {
     //  Launch PostgreSQL server. Download / unarchive if necessary.
-    postgres.start(EmbeddedPostgres.cachedRuntimeConfig(Paths.get("target/postgres")), "localhost", 5432, "refcursor", "postgres", "root", Collections.emptyList());
+    postgres.start(EmbeddedPostgres.cachedRuntimeConfig(Paths.get("target/postgres")), "localhost", 5432, "keycolumn", "postgres", "root", Collections.emptyList());
 
-    Reader reader = Resources.getResourceAsReader("org/apache/ibatis/submitted/refcursor/MapperConfig.xml");
+    Reader reader = Resources.getResourceAsReader("org/apache/ibatis/submitted/usesjava8/keycolumn/MapperConfig.xml");
     sqlSessionFactory = new SqlSessionFactoryBuilder().build(reader);
     reader.close();
 
     SqlSession session = sqlSessionFactory.openSession();
     Connection conn = session.getConnection();
-    reader = Resources.getResourceAsReader("org/apache/ibatis/submitted/refcursor/CreateDB.sql");
+    reader = Resources.getResourceAsReader("org/apache/ibatis/submitted/usesjava8/keycolumn/CreateDB.sql");
     ScriptRunner runner = new ScriptRunner(conn);
     runner.setLogWriter(null);
     runner.runScript(reader);
@@ -73,42 +73,62 @@ public class RefCursorTest {
   }
 
   @Test
-  public void testRefCursor1() throws IOException {
+  public void testInsertAnnotated() throws Exception {
     SqlSession sqlSession = sqlSessionFactory.openSession();
     try {
-      OrdersMapper mapper = sqlSession.getMapper(OrdersMapper.class);
-      Map<String, Object> parameter = new HashMap<String, Object>();
-      parameter.put("orderId", 1);
-      mapper.getOrder1(parameter);
+      InsertMapper mapper = sqlSession.getMapper(InsertMapper.class);
+      Name name = new Name();
+      name.setFirstName("Fred");
+      name.setLastName("Flintstone");
 
-      assertNotNull(parameter.get("order"));
-      @SuppressWarnings("unchecked")
-      List<Order> orders = (List<Order>) parameter.get("order");
-      assertEquals(1, orders.size());
-      Order order = orders.get(0);
-      assertEquals(3, order.getDetailLines().size());
+      int rows = mapper.insertNameAnnotated(name);
+
+      assertNotNull(name.getId());
+      assertEquals(1, rows);
     } finally {
       sqlSession.close();
     }
   }
 
   @Test
-  public void testRefCursor2() throws IOException {
+  public void testInsertMapped() throws Exception {
     SqlSession sqlSession = sqlSessionFactory.openSession();
     try {
-      OrdersMapper mapper = sqlSession.getMapper(OrdersMapper.class);
-      Map<String, Object> parameter = new HashMap<String, Object>();
-      parameter.put("orderId", 1);
-      mapper.getOrder2(parameter);
+      InsertMapper mapper = sqlSession.getMapper(InsertMapper.class);
+      Name name = new Name();
+      name.setFirstName("Fred");
+      name.setLastName("Flintstone");
 
-      assertNotNull(parameter.get("order"));
-      @SuppressWarnings("unchecked")
-      List<Order> orders = (List<Order>) parameter.get("order");
-      assertEquals(1, orders.size());
-      Order order = orders.get(0);
-      assertEquals(3, order.getDetailLines().size());
+      int rows = mapper.insertNameMapped(name);
+
+      assertNotNull(name.getId());
+      assertEquals(1, rows);
     } finally {
       sqlSession.close();
     }
   }
+
+  @Ignore // Not supported yet in PostgreSQL
+  @Test
+  public void testInsertMappedBatch() throws Exception {
+    SqlSession sqlSession = sqlSessionFactory.openSession(ExecutorType.BATCH);
+    try {
+      InsertMapper mapper = sqlSession.getMapper(InsertMapper.class);
+      Name name = new Name();
+      name.setFirstName("Fred");
+      name.setLastName("Flintstone");
+      mapper.insertNameMapped(name);
+      Name name2 = new Name();
+      name2.setFirstName("Wilma");
+      name2.setLastName("Flintstone");
+      mapper.insertNameMapped(name2);
+      List<BatchResult> batchResults = sqlSession.flushStatements();
+      assertNotNull(name.getId());
+      assertNotNull(name2.getId());
+      assertEquals(1, batchResults.size());
+    } finally {
+      sqlSession.close();
+    }
+  }
+
 }
