+++ src/test/java/org/apache/ibatis/submitted/usesjava8/keycolumn/Name.java
@@ -13,12 +13,28 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.apache.ibatis.submitted.multiple_resultsets;
+package org.apache.ibatis.submitted.usesjava8.keycolumn;
 
-import java.util.List;
-
-public interface Mapper {
-
-  List<?> getUsersAndGroups();
-
+public class Name {
+    private Integer id;
+    private String firstName;
+    private String lastName;
+    public Integer getId() {
+        return id;
+    }
+    public void setId(Integer id) {
+        this.id = id;
+    }
+    public String getFirstName() {
+        return firstName;
+    }
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+    public String getLastName() {
+        return lastName;
+    }
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
 }
