+++ src/test/java/org/apache/ibatis/submitted/usesjava8/multiple_resultsets/Mapper.java
@@ -13,28 +13,12 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.apache.ibatis.submitted.keycolumn;
+package org.apache.ibatis.submitted.usesjava8.multiple_resultsets;
 
-public class Name {
-    private Integer id;
-    private String firstName;
-    private String lastName;
-    public Integer getId() {
-        return id;
-    }
-    public void setId(Integer id) {
-        this.id = id;
-    }
-    public String getFirstName() {
-        return firstName;
-    }
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-    public String getLastName() {
-        return lastName;
-    }
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
+import java.util.List;
+
+public interface Mapper {
+
+  List<?> getUsersAndGroups();
+
 }
