+++ src/test/java/org/apache/ibatis/submitted/usesjava8/refcursor/OrdersMapper.java
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.apache.ibatis.submitted.refcursor;
+package org.apache.ibatis.submitted.usesjava8.refcursor;
 
 import java.util.Map;
