+++ src/test/java/org/apache/ibatis/submitted/usesjava8/refcursor/OrdersMapper.xml
@@ -17,11 +17,11 @@
 
 -->
 <!DOCTYPE mapper PUBLIC "-//ibatis.apache.org//DTD Mapper 3.0//EN" "http://ibatis.apache.org/dtd/ibatis-3-mapper.dtd">
-<mapper namespace="org.apache.ibatis.submitted.refcursor.OrdersMapper">
-  <resultMap type="org.apache.ibatis.submitted.refcursor.Order" id="OrderResult">
+<mapper namespace="org.apache.ibatis.submitted.usesjava8.refcursor.OrdersMapper">
+  <resultMap type="org.apache.ibatis.submitted.usesjava8.refcursor.Order" id="OrderResult">
     <id property="orderId" column="ORDER_ID"/>
     <result property="customerName" column="CUST_NAME"/>
-    <collection property="detailLines" ofType="org.apache.ibatis.submitted.refcursor.OrderDetail">
+    <collection property="detailLines" ofType="org.apache.ibatis.submitted.usesjava8.refcursor.OrderDetail">
       <id property="lineNumber" column="LINE_NUMBER"/>
       <result property="quantity" column="QUANTITY"/>
       <result property="description" column="ITEM_DESCRIPTION"/>
@@ -34,7 +34,7 @@
   </update>
 
   <update id="getOrder2" parameterType="map" statementType="CALLABLE">
-    { #{order,jdbcType=OTHER,mode=OUT,resultMap=org.apache.ibatis.submitted.refcursor.OrdersMapper.OrderResult,javaType=java.sql.ResultSet} = 
+    { #{order,jdbcType=OTHER,mode=OUT,resultMap=org.apache.ibatis.submitted.usesjava8.refcursor.OrdersMapper.OrderResult,javaType=java.sql.ResultSet} = 
         call mbtest.get_order(#{orderId,jdbcType=INTEGER,mode=IN}) }
   </update>
 </mapper>
